several modification after hackaton
