+++ src/automl/model/blender.py
@@ -248,14 +248,14 @@ class CoordDescBlender(BaseModel):
         weights[idx] = weight
         weights = self.adjust_weights(weights, idx, self.drop_thresh)
         y_pred = self._compute_weighted_pred(X, weights)
-        
+
         # not_nan_idx = ~np.isnan(y_pred).any(axis=1)
         # y_pred = y_pred[not_nan_idx]
 
         if y_pred.ndim == 2 and y_pred.shape[1] == 2:
             # binary case
             y_pred = y_pred[:, 1]
-            
+
         metric = scorer.score(y, y_pred)
         return -1 * metric if scorer.greater_is_better else metric
