+++ src/automl/model/xgboost/xgboost.py
@@ -297,11 +297,11 @@ class XGBClassification(BaseModel):
         self.time_series = time_series
         self.eval_metric = eval_metric
         self.device = device
-        
+
         if self.device is None and torch.cuda.is_available():
-            self.device="cuda"
+            self.device = "cuda"
         else:
-            self.device="cpu"
+            self.device = "cpu"
 
         if self.time_series:
             self.kf = TimeSeriesSplit(n_splits=5)
@@ -309,9 +309,9 @@ class XGBClassification(BaseModel):
             self.kf = StratifiedKFold(
                 n_splits=5, random_state=self.random_state, shuffle=True
             )
-        
-        self.models = None
-        self.oof_preds = None
+
+        # self.models = None
+        # self.oof_preds = None
 
     def fit(self, X: FeaturesType, y: TargetType, categorical_features=[]):
         log.info(f"Fitting {self.name}", msg_type="start")
@@ -451,24 +451,24 @@ class XGBClassification(BaseModel):
             trial_metric = scorer.score(y[not_none_oof], oof_preds[not_none_oof, 1])
         else:
             trial_metric = scorer.score(y[not_none_oof], oof_preds[not_none_oof])
-        
-        if self.models is None:
-            # no trials are completed yet
-            # save tuned models
-            self.models = models
-            self.oof_preds = oof_preds
-            
-        elif (trial_metric <= trial.study.best_value and trial.study.direction == 1) or (
-            trial_metric >= trial.study.best_value and trial.study.direction == 2
-            ):
-            
-            # new best 
-            # save tuned models
-            self.models = models
-            self.oof_preds = oof_preds
-            
+
+        # if self.models is None:
+        #     # no trials are completed yet
+        #     # save tuned models
+        #     self.models = models
+        #     self.oof_preds = oof_preds
+
+        # elif (trial_metric <= trial.study.best_value and trial.study.direction == 1) or (
+        #     trial_metric >= trial.study.best_value and trial.study.direction == 2
+        #     ):
+
+        #     # new best
+        #     # save tuned models
+        #     self.models = models
+        #     self.oof_preds = oof_preds
+
         return trial_metric
-    
+
     def tune(
         self,
         X: FeaturesType,
